+++ modules/core/src/main/java/org/apache/ignite/internal/management/ChangeTagCommand.java
@@ -21,7 +21,6 @@ import java.util.function.Consumer;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 import org.apache.ignite.internal.visor.misc.VisorClusterChangeTagTask;
 import org.apache.ignite.internal.visor.misc.VisorClusterChangeTagTaskResult;
-import org.jetbrains.annotations.Nullable;
 
 /** Change Cluster tag command. */
 public class ChangeTagCommand implements ComputeCommand<ChangeTagCommandArg, VisorClusterChangeTagTaskResult> {
@@ -53,7 +52,7 @@ public class ChangeTagCommand implements ComputeCommand<ChangeTagCommandArg, Vis
     }
 
     /** {@inheritDoc} */
-    @Override public @Nullable String confirmationPrompt(ChangeTagCommandArg arg) {
+    @Override public String confirmationPrompt(ChangeTagCommandArg arg) {
         return "Warning: the command will change cluster tag.";
     }
 }
