+++ modules/core/src/main/java/org/apache/ignite/internal/management/meta/MetaUpdateCommand.java
@@ -23,7 +23,6 @@ import org.apache.ignite.internal.commandline.meta.tasks.MetadataMarshalled;
 import org.apache.ignite.internal.commandline.meta.tasks.MetadataUpdateTask;
 import org.apache.ignite.internal.management.api.ComputeCommand;
 import org.apache.ignite.internal.management.api.ExperimentalCommand;
-import org.jetbrains.annotations.Nullable;
 
 /** */
 public class MetaUpdateCommand implements
@@ -58,7 +57,7 @@ public class MetaUpdateCommand implements
     }
 
     /** {@inheritDoc} */
-    @Override public @Nullable String confirmationPrompt(MetaUpdateCommandArg arg) {
+    @Override public String confirmationPrompt(MetaUpdateCommandArg arg) {
         return "Warning: the command will update the binary metadata at the cluster.";
     }
 }
